pyproject.toml
[project]
name = "transitflow"
version = "0.1.0"
description = "Real-time delay prediction for public transit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core logic and data validation
    "paho-mqtt>=2.0.0",
    "confluent-kafka>=2.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.1",
    "fastavro>=1.9.0",
    
    # Data Processing and Storage
    "redis>=5.0.0",
    "pyspark==3.5.0",
    "delta-spark==3.0.0",
    
    # Warehouse and Transformation
    "dbt-core~=1.8.0",
    "dbt-postgres~=1.8.0",
    "psycopg2-binary>=2.9.9", 
    "python-dotenv>=1.0.1",

    # Feature Store & Serving API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",

    # Machine Learning & Analysis
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "mlflow==2.10.2",
    "boto3>=1.34.0,<1.42.31",
    
    # Observability & Tracing
    "prometheus-client>=0.19.0",
    "structlog>=24.0.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    # Required for Python 3.12+ because pkg_resources was removed from stdlib
    "setuptools<80.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "faker>=22.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  # Look in the root directory
include = ["ingestion*", "schema_registry*", "feature_store*", "serving*", "ml_pipeline*"]
exclude = ["tests*", "scripts*", "infra*", "docker*", "ui*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
# Moved select/ignore to the correct modern section
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.bandit]
# Security scanning configuration
exclude_dirs = ["tests", ".venv"]
skips = ["B101"] # B101 ignores 'assert' statements